requirements.txt
dashscope>=1.14.0
flask>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
